lowering: move var_id and LiveRange into live_variable_analysis
